Add helper module to assert required and default opts
